improve pipeline, need to fix train, test, process, summary.
