--- src/data.py
@@ -1,83 +1,0 @@
-import os
-import numpy as np
-import datasets
-from config import cfg
-from torchvision import transforms
-from torch.utils.data import DataLoader, Dataset
-from torch.utils.data.dataloader import default_collate
-
-data_stats = {'MNIST': ((0.1307,), (0.3081,)), 'FashionMNIST': ((0.2860,), (0.3530,)),
-              'CIFAR10': ((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
-              'CIFAR100': ((0.5071, 0.4865, 0.4409), (0.2673, 0.2564, 0.2762)),
-              'SVHN': ((0.4377, 0.4438, 0.4728), (0.1980, 0.2010, 0.1970))}
-
-
-def fetch_dataset(data_name, verbose=True):
-    dataset = {}
-    if verbose:
-        print('fetching data {}...'.format(data_name))
-    root = os.path.join('data', data_name)
-    if data_name in ['MNIST', 'FashionMNIST']:
-        dataset['train'] = eval('datasets.{}(root=root, split="train", '
-                                'transform=datasets.Compose([transforms.ToTensor()]))'.format(data_name))
-        dataset['test'] = eval('datasets.{}(root=root, split="test", '
-                               'transform=datasets.Compose([transforms.ToTensor()]))'.format(data_name))
-        dataset['train'].transform = datasets.Compose([
-            transforms.ToTensor(),
-            transforms.Normalize(*data_stats[data_name])])
-        dataset['test'].transform = datasets.Compose([
-            transforms.ToTensor(),
-            transforms.Normalize(*data_stats[data_name])])
-    elif data_name in ['CIFAR10', 'CIFAR100']:
-        dataset['train'] = eval('datasets.{}(root=root, split="train", '
-                                'transform=datasets.Compose([transforms.ToTensor()]))'.format(data_name))
-        dataset['test'] = eval('datasets.{}(root=root, split="test", '
-                               'transform=datasets.Compose([transforms.ToTensor()]))'.format(data_name))
-        dataset['train'].transform = datasets.Compose([
-            transforms.RandomHorizontalFlip(),
-            transforms.RandomCrop(32, padding=4, padding_mode='reflect'),
-            transforms.ToTensor(),
-            transforms.Normalize(*data_stats[data_name])])
-        dataset['test'].transform = datasets.Compose([
-            transforms.ToTensor(),
-            transforms.Normalize(*data_stats[data_name])])
-    elif data_name in ['SVHN']:
-        dataset['train'] = eval('datasets.{}(root=root, split="train", '
-                                'transform=datasets.Compose([transforms.ToTensor()]))'.format(data_name))
-        dataset['test'] = eval('datasets.{}(root=root, split="test", '
-                               'transform=datasets.Compose([transforms.ToTensor()]))'.format(data_name))
-        dataset['train'].transform = datasets.Compose([
-            transforms.RandomCrop(32, padding=4, padding_mode='reflect'),
-            transforms.ToTensor(),
-            transforms.Normalize(*data_stats[data_name])])
-        dataset['test'].transform = datasets.Compose([
-            transforms.ToTensor(),
-            transforms.Normalize(*data_stats[data_name])])
-    else:
-        raise ValueError('Not valid dataset name')
-    if verbose:
-        print('data ready')
-    return dataset
-
-
-def input_collate(batch):
-    if isinstance(batch[0], dict):
-        return {key: [b[key] for b in batch] for key in batch[0]}
-    else:
-        return default_collate(batch)
-
-
-def make_data_loader(dataset, tag, batch_size=None, shuffle=None, sampler=None):
-    data_loader = {}
-    for k in dataset:
-        _batch_size = cfg[tag]['batch_size'][k] if batch_size is None else batch_size[k]
-        _shuffle = cfg[tag]['shuffle'][k] if shuffle is None else shuffle[k]
-        if sampler is None:
-            data_loader[k] = DataLoader(dataset=dataset[k], batch_size=_batch_size, shuffle=_shuffle,
-                                        pin_memory=True, num_workers=cfg['num_workers'], collate_fn=input_collate,
-                                        worker_init_fn=np.random.seed(cfg['seed']))
-        else:
-            data_loader[k] = DataLoader(dataset=dataset[k], batch_size=_batch_size, sampler=sampler[k],
-                                        pin_memory=True, num_workers=cfg['num_workers'], collate_fn=input_collate,
-                                        worker_init_fn=np.random.seed(cfg['seed']))
-    return data_loader
