+++ src/dataset/__init__.py
@@ -1,0 +1,5 @@
+from .dataset import *
+from .utils import *
+from .mnist import MNIST, FashionMNIST
+from .cifar import CIFAR10, CIFAR100
+from .svhn import SVHN
