+++ src/dataset/dataset.py
@@ -1,0 +1,118 @@
+import dataset
+import numpy as np
+import os
+import torch
+from torchvision import transforms
+from torch.utils.data import DataLoader
+from torch.utils.data.dataloader import default_collate
+from config import cfg
+
+data_stats = {'MNIST': ((0.1307,), (0.3081,)), 'FashionMNIST': ((0.2860,), (0.3530,)),
+              'CIFAR10': ((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
+              'CIFAR100': ((0.5071, 0.4865, 0.4409), (0.2673, 0.2564, 0.2762)),
+              'SVHN': ((0.4377, 0.4438, 0.4728), (0.1980, 0.2010, 0.1970))}
+
+
+def make_dataset(data_name, verbose=True):
+    dataset_ = {}
+    if verbose:
+        print('fetching data {}...'.format(data_name))
+    root = os.path.join('data', data_name)
+    if data_name in ['MNIST', 'FashionMNIST']:
+        dataset_['train'] = eval('dataset.{}(root=root, split="train", '
+                                 'transform=dataset.Compose([transforms.ToTensor()]))'.format(data_name))
+        dataset_['test'] = eval('dataset.{}(root=root, split="test", '
+                                'transform=dataset.Compose([transforms.ToTensor()]))'.format(data_name))
+        dataset_['train'].transform = dataset.Compose([
+            transforms.ToTensor(),
+            transforms.Normalize(*data_stats[data_name])])
+        dataset_['test'].transform = dataset.Compose([
+            transforms.ToTensor(),
+            transforms.Normalize(*data_stats[data_name])])
+    elif data_name in ['CIFAR10', 'CIFAR100']:
+        dataset_['train'] = eval('dataset.{}(root=root, split="train", '
+                                 'transform=dataset.Compose([transforms.ToTensor()]))'.format(data_name))
+        dataset_['test'] = eval('dataset.{}(root=root, split="test", '
+                                'transform=dataset.Compose([transforms.ToTensor()]))'.format(data_name))
+        dataset_['train'].transform = dataset.Compose([
+            transforms.RandomHorizontalFlip(),
+            transforms.RandomCrop(32, padding=4, padding_mode='reflect'),
+            transforms.ToTensor(),
+            transforms.Normalize(*data_stats[data_name])])
+        dataset_['test'].transform = dataset.Compose([
+            transforms.ToTensor(),
+            transforms.Normalize(*data_stats[data_name])])
+    elif data_name in ['SVHN']:
+        dataset_['train'] = eval('dataset.{}(root=root, split="train", '
+                                 'transform=dataset.Compose([transforms.ToTensor()]))'.format(data_name))
+        dataset_['test'] = eval('dataset.{}(root=root, split="test", '
+                                'transform=dataset.Compose([transforms.ToTensor()]))'.format(data_name))
+        dataset_['train'].transform = dataset.Compose([
+            transforms.RandomCrop(32, padding=4, padding_mode='reflect'),
+            transforms.ToTensor(),
+            transforms.Normalize(*data_stats[data_name])])
+        dataset_['test'].transform = dataset.Compose([
+            transforms.ToTensor(),
+            transforms.Normalize(*data_stats[data_name])])
+    elif data_name in ['FPB']:
+        dataset_ = load_dataset('financial_phrasebank', 'sentences_allagree', cache_dir=root)
+        dataset_ = dataset_['train'].train_test_split(test_size=0.1)
+        classes = dataset_['train'].features['label'].names
+        dataset_ = dataset_.map(
+            lambda x: {"text_label": [classes[label] for label in x["label"]]},
+            batched=True,
+            num_proc=1,
+        )
+    else:
+        raise ValueError('Not valid dataset name')
+    if verbose:
+        print('data ready')
+    return dataset_
+
+
+def input_collate(batch):
+    return {key: [b[key] for b in batch] for key in batch[0]}
+
+
+def make_data_collate(collate_mode):
+    if collate_mode == 'dict':
+        return input_collate
+    elif collate_mode == 'default':
+        return default_collate
+    elif collate_mode == 'transformer':
+        return default_data_collator
+    else:
+        raise ValueError('Not valid collate mode')
+
+
+def make_data_loader(dataset, tag, batch_size=None, shuffle=None, sampler=None):
+    data_loader = {}
+    cfg['num_steps'] = {}
+    for k in dataset:
+        _batch_size = cfg[tag]['batch_size'][k] if batch_size is None else batch_size[k]
+        _shuffle = cfg[tag]['shuffle'][k] if shuffle is None else shuffle[k]
+        if sampler is None:
+            data_loader[k] = DataLoader(dataset=dataset[k], batch_size=_batch_size, shuffle=_shuffle,
+                                        pin_memory=cfg['pin_memory'], num_workers=cfg['num_workers'],
+                                        collate_fn=make_data_collate(cfg['collate_mode']),
+                                        worker_init_fn=np.random.seed(cfg['seed']))
+        else:
+            data_loader[k] = DataLoader(dataset=dataset[k], batch_size=_batch_size, sampler=sampler[k],
+                                        pin_memory=cfg['pin_memory'], num_workers=cfg['num_workers'],
+                                        collate_fn=make_data_collate(cfg['collate_mode']),
+                                        worker_init_fn=np.random.seed(cfg['seed']))
+        cfg['num_steps'][k] = len(data_loader[k])
+    return data_loader
+
+
+def collate(input):
+    for k in input:
+        input[k] = torch.stack(input[k], 0)
+    return input
+
+
+def process_dataset(dataset):
+    processed_dataset = dataset
+    cfg['data_size'] = {'train': len(processed_dataset['train']), 'test': len(processed_dataset['test'])}
+    cfg['target_size'] = dataset['train'].target_size
+    return processed_dataset
