+++ src/dataset/mnist.py
@@ -4,7 +4,7 @@ import os
 import torch
 from PIL import Image
 from torch.utils.data import Dataset
-from utils import check_exists, makedir_exist_ok, save, load
+from module import check_exists, makedir_exist_ok, save, load
 from .utils import download_url, extract_file, make_classes_counts
