--- src/datasets/__init__.py
@@ -1,6 +1,0 @@
-from .mnist import MNIST, FashionMNIST
-from .cifar import CIFAR10, CIFAR100
-from .svhn import SVHN
-from .utils import *
-
-__all__ = ('MNIST', 'FashionMNIST', 'CIFAR10', 'CIFAR100', 'SVHN')
