+++ src/make_stats.py
@@ -1,21 +1,20 @@
 import os
 import torch
-import datasets
 from torchvision import transforms
 from config import cfg
-from data import fetch_dataset, make_data_loader
-from utils import save, collate, Stats, makedir_exist_ok, process_control, process_dataset
+from dataset import make_dataset, make_data_loader, process_dataset, collate, Compose
+from module import save, Stats, makedir_exist_ok, process_control
 
 if __name__ == "__main__":
-    stats_path = os.path.join('res', 'stats')
+    stats_path = os.path.join('output', 'stats')
     dim = 1
     data_names = ['MNIST', 'FashionMNIST', 'SVHN', 'CIFAR10', 'CIFAR100']
     process_control()
     cfg['seed'] = 0
     with torch.no_grad():
         for data_name in data_names:
-            dataset = fetch_dataset(data_name)
-            dataset['train'].transform = datasets.Compose([transforms.ToTensor()])
+            dataset = make_dataset(data_name)
+            dataset['train'].transform = Compose([transforms.ToTensor()])
             process_dataset(dataset)
             data_loader = make_data_loader(dataset, cfg['model_name'])
             stats = Stats(dim=dim)
@@ -25,4 +24,4 @@ if __name__ == "__main__":
             stats = (stats.mean.tolist(), stats.std.tolist())
             print(data_name, stats)
             makedir_exist_ok(stats_path)
-            save(stats, os.path.join(stats_path, '{}.pt'.format(data_name)))
+            save(stats, os.path.join(stats_path, '{}'.format(data_name)))
