--- src/models/__init__.py
@@ -1,6 +1,0 @@
-from .utils import *
-from .linear import *
-from .mlp import *
-from .cnn import *
-from .resnet import *
-from .wresnet import *
