--- src/models/utils.py
@@ -1,95 +1,0 @@
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from config import cfg
-
-
-def init_param(m):
-    if isinstance(m, (nn.Conv2d, nn.ConvTranspose2d)):
-        nn.init.kaiming_normal_(m.weight, mode='fan_out', nonlinearity='relu')
-        if m.bias is not None:
-            m.bias.data.zero_()
-    elif isinstance(m, nn.BatchNorm2d):
-        if m.weight is not None:
-            m.weight.data.fill_(1)
-        if m.bias is not None:
-            m.bias.data.zero_()
-    elif isinstance(m, nn.Linear):
-        if m.bias is not None:
-            m.bias.data.zero_()
-    return m
-
-
-def init_param_generator(m):
-    if isinstance(m, (nn.Conv2d, nn.ConvTranspose2d)):
-        nn.init.normal_(m.weight.data, 0.0, 0.02)
-    elif isinstance(m, nn.BatchNorm2d):
-        if m.weight is not None:
-            m.weight.data.fill_(1)
-        if m.bias is not None:
-            m.bias.data.zero_()
-    elif isinstance(m, nn.Linear):
-        nn.init.normal_(m.weight.data, 0.0, 0.02)
-    return m
-
-
-def normalize(input):
-    if cfg['data_name'] in cfg['stats']:
-        broadcast_size = [1] * input.dim()
-        broadcast_size[1] = input.size(1)
-        m, s = cfg['stats'][cfg['data_name']]
-        m, s = torch.tensor(m, dtype=input.dtype).view(broadcast_size).to(input.device), \
-               torch.tensor(s, dtype=input.dtype).view(broadcast_size).to(input.device)
-        input = input.sub(m).div(s)
-    return input
-
-
-def denormalize(input):
-    if cfg['data_name'] in cfg['stats']:
-        broadcast_size = [1] * input.dim()
-        broadcast_size[1] = input.size(1)
-        m, s = cfg['stats'][cfg['data_name']]
-        m, s = torch.tensor(m, dtype=input.dtype).view(broadcast_size).to(input.device), \
-               torch.tensor(s, dtype=input.dtype).view(broadcast_size).to(input.device)
-        input = input.mul(s).add(m)
-    return input
-
-
-def make_batchnorm(m, momentum, track_running_stats):
-    if isinstance(m, (nn.BatchNorm1d, nn.BatchNorm2d)):
-        m.momentum = momentum
-        m.track_running_stats = track_running_stats
-        if track_running_stats:
-            m.register_buffer('running_mean', torch.zeros(m.num_features, device=cfg['device']))
-            m.register_buffer('running_var', torch.ones(m.num_features, device=cfg['device']))
-            m.register_buffer('num_batches_tracked', torch.tensor(0, dtype=torch.long, device=cfg['device']))
-        else:
-            m.running_mean = None
-            m.running_var = None
-            m.num_batches_tracked = None
-    return m
-
-
-def make_loss(output, input):
-    loss = loss_fn(output['target'], input['target'])
-    return loss
-
-
-def loss_fn(output, target, reduction='mean'):
-    if target.dtype == torch.int64:
-        loss = F.cross_entropy(output, target, reduction=reduction)
-    else:
-        loss = kld_loss(output, target, reduction=reduction)
-    return loss
-
-
-def cross_entropy_loss(output, target, reduction='mean'):
-    if target.dtype != torch.int64:
-        target = (target.topk(1, 1, True, True)[1]).view(-1)
-    ce = F.cross_entropy(output, target, reduction=reduction)
-    return ce
-
-
-def kld_loss(output, target, reduction='batchmean'):
-    kld = F.kl_div(F.log_softmax(output, dim=-1), target, reduction=reduction)
-    return kld
