--- src/models/wresnet.py
@@ -1,117 +1,0 @@
-import math
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from .utils import init_param, make_loss
-from config import cfg
-
-
-class BasicBlock(nn.Module):
-    def __init__(self, in_planes, out_planes, stride, drop_rate):
-        super().__init__()
-        self.n1 = nn.BatchNorm2d(in_planes)
-        self.relu1 = nn.ReLU(inplace=True)
-        self.conv1 = nn.Conv2d(in_planes, out_planes, kernel_size=3, stride=stride,
-                               padding=1, bias=False)
-        self.n2 = nn.BatchNorm2d(out_planes)
-        self.relu2 = nn.ReLU(inplace=True)
-        self.conv2 = nn.Conv2d(out_planes, out_planes, kernel_size=3, stride=1,
-                               padding=1, bias=False)
-        self.drop_rate = drop_rate
-        self.equal_inout = (in_planes == out_planes)
-        self.shortcut = (not self.equal_inout) and nn.Conv2d(in_planes, out_planes, kernel_size=1, stride=stride,
-                                                             padding=0, bias=False) or None
-
-    def forward(self, x):
-        if not self.equal_inout:
-            x = self.relu1(self.n1(x))
-        else:
-            out = self.relu1(self.n1(x))
-        out = self.relu2(self.n2(self.conv1(out if self.equal_inout else x)))
-        if self.drop_rate > 0:
-            out = F.dropout(out, p=self.drop_rate, training=self.training)
-        out = self.conv2(out)
-        out = torch.add(x if self.equal_inout else self.shortcut(x), out)
-        return out
-
-
-class NetworkBlock(nn.Module):
-    def __init__(self, nb_layers, in_planes, out_planes, block, stride, drop_rate):
-        super().__init__()
-        self.layer = self._make_layer(block, in_planes, out_planes, nb_layers, stride, drop_rate)
-
-    def _make_layer(self, block, in_planes, out_planes, nb_layers, stride, drop_rate):
-        layers = []
-        for i in range(int(nb_layers)):
-            layers.append(block(i == 0 and in_planes or out_planes, out_planes, i == 0 and stride or 1, drop_rate))
-        return nn.Sequential(*layers)
-
-    def forward(self, x):
-        x = self.layer(x)
-        return x
-
-
-class WideResNet(nn.Module):
-    def __init__(self, data_shape, num_classes, depth, widen_factor, drop_rate):
-        super().__init__()
-        num_down = int(min(round(math.log2(data_shape[1])), round(math.log2(data_shape[2])))) - 3
-        hidden_size = [16]
-        for i in range(num_down + 1):
-            hidden_size.append(16 * (2 ** i) * widen_factor)
-        self.hidden_size = hidden_size
-        n = ((depth - 1) / (num_down + 1) - 1) / 2
-        block = BasicBlock
-        blocks = []
-        blocks.append(nn.Conv2d(data_shape[0], hidden_size[0], kernel_size=3, stride=1, padding=1, bias=False))
-        blocks.append(NetworkBlock(n, hidden_size[0], hidden_size[1], block, 1, drop_rate))
-        for i in range(num_down):
-            blocks.append(NetworkBlock(n, hidden_size[i + 1], hidden_size[i + 2], block, 2, drop_rate))
-        blocks.append(nn.BatchNorm2d(hidden_size[-1]))
-        blocks.append(nn.ReLU(inplace=True))
-        blocks.append(nn.AdaptiveAvgPool2d(1))
-        blocks.append(nn.Flatten())
-        self.blocks = nn.Sequential(*blocks)
-        self.linear = nn.Linear(hidden_size[-1], num_classes)
-
-    def feature(self, x):
-        x = self.blocks(x)
-        return x
-
-    def classify(self, x):
-        x = self.linear(x)
-        return x
-
-    def f(self, x):
-        x = self.feature(x)
-        x = self.classify(x)
-        return x
-
-    def forward(self, input):
-        output = {}
-        x = input['data']
-        x = self.f(x)
-        output['target'] = x
-        output['loss'] = make_loss(output, input)
-        return output
-
-
-def wresnet28x2():
-    data_shape = cfg['data_shape']
-    target_size = cfg['target_size']
-    depth = cfg['wresnet28x2']['depth']
-    widen_factor = cfg['wresnet28x2']['widen_factor']
-    drop_rate = cfg['wresnet28x2']['drop_rate']
-    model = WideResNet(data_shape, target_size, depth, widen_factor, drop_rate)
-    model.apply(init_param)
-    return model
-
-
-def wresnet28x8():
-    data_shape = cfg['data_shape']
-    target_size = cfg['target_size']
-    depth = cfg['wresnet28x8']['depth']
-    widen_factor = cfg['wresnet28x8']['widen_factor']
-    drop_rate = cfg['wresnet28x8']['drop_rate']
-    model = WideResNet(data_shape, target_size, depth, widen_factor, drop_rate)
-    model.apply(init_param)
-    return model
