+++ src/module/__init__.py
@@ -1,0 +1,4 @@
+from .utils import *
+from .hyper import *
+from .io import *
+from .stats import *
