+++ src/module/stats.py
@@ -1,0 +1,43 @@
+import os
+import torch
+from .io import makedir_exist_ok, load
+
+
+def make_stats():
+    stats = {}
+    stats_path = os.path.join('output', 'res')
+    makedir_exist_ok(stats_path)
+    filenames = os.listdir(stats_path)
+    for filename in filenames:
+        stats_name = os.path.splitext(filename)[0]
+        stats[stats_name] = load(os.path.join(stats_path, filename))
+    return stats
+
+
+class Stats(object):
+    def __init__(self, dim):
+        self.dim = dim
+        self.n_samples = 0
+        self.n_features = None
+        self.mean = None
+        self.std = None
+
+    def update(self, data):
+        data = data.transpose(self.dim, -1).reshape(-1, data.size(self.dim))
+        if self.n_samples == 0:
+            self.n_samples = data.size(0)
+            self.n_features = data.size(1)
+            self.mean = data.mean(dim=0)
+            self.std = data.std(dim=0)
+        else:
+            m = float(self.n_samples)
+            n = data.size(0)
+            new_mean = data.mean(dim=0)
+            new_std = 0 if n == 1 else data.std(dim=0)
+            old_mean = self.mean
+            old_std = self.std
+            self.mean = m / (m + n) * old_mean + n / (m + n) * new_mean
+            self.std = torch.sqrt(m / (m + n) * old_std ** 2 + n / (m + n) * new_std ** 2 + m * n / (m + n) ** 2 * (
+                    old_mean - new_mean) ** 2)
+            self.n_samples += n
+        return
