+++ src/module/utils.py
@@ -1,0 +1,38 @@
+import numpy as np
+import torch
+import collections.abc as container_abcs
+from itertools import repeat
+
+
+def ntuple(n):
+    def parse(x):
+        if isinstance(x, container_abcs.Iterable) and not isinstance(x, str):
+            return x
+        return tuple(repeat(x, n))
+
+    return parse
+
+
+def recur(fn, input, *args):
+    if isinstance(input, torch.Tensor) or isinstance(input, np.ndarray):
+        output = fn(input, *args)
+    elif isinstance(input, list):
+        output = []
+        for i in range(len(input)):
+            output.append(recur(fn, input[i], *args))
+    elif isinstance(input, tuple):
+        output = []
+        for i in range(len(input)):
+            output.append(recur(fn, input[i], *args))
+        output = tuple(output)
+    elif isinstance(input, dict):
+        output = {}
+        for key in input:
+            output[key] = recur(fn, input[key], *args)
+    elif isinstance(input, str):
+        output = input
+    elif input is None:
+        output = None
+    else:
+        raise ValueError('Not valid input type')
+    return output
