--- src/test.py
@@ -1,10 +1,0 @@
-import argparse
-import os
-import torch
-import torch.backends.cudnn as cudnn
-import models
-from config import cfg, process_args
-from data import fetch_dataset, make_data_loader
-from metrics import Metric
-from utils import save, to_device, process_control, process_dataset, resume, collate
-from logger import make_logger
