--- src/utils.py
@@ -1,234 +1,0 @@
-import collections.abc as container_abcs
-import errno
-import numpy as np
-import os
-import pickle
-import torch
-import torch.optim as optim
-from itertools import repeat
-from torchvision.utils import save_image
-from config import cfg
-
-
-def check_exists(path):
-    return os.path.exists(path)
-
-
-def makedir_exist_ok(path):
-    try:
-        os.makedirs(path)
-    except OSError as e:
-        if e.errno == errno.EEXIST:
-            pass
-        else:
-            raise
-    return
-
-
-def save(input, path, mode='torch'):
-    dirname = os.path.dirname(path)
-    makedir_exist_ok(dirname)
-    if mode == 'torch':
-        torch.save(input, path)
-    elif mode == 'np':
-        np.save(path, input, allow_pickle=True)
-    elif mode == 'pickle':
-        pickle.dump(input, open(path, 'wb'))
-    else:
-        raise ValueError('Not valid save mode')
-    return
-
-
-def load(path, mode='torch'):
-    if mode == 'torch':
-        return torch.load(path, map_location=lambda storage, loc: storage)
-    elif mode == 'np':
-        return np.load(path, allow_pickle=True)
-    elif mode == 'pickle':
-        return pickle.load(open(path, 'rb'))
-    else:
-        raise ValueError('Not valid save mode')
-    return
-
-
-def save_img(img, path, nrow=10, padding=1, pad_value=0, value_range=None):
-    makedir_exist_ok(os.path.dirname(path))
-    normalize = False if range is None else True
-    save_image(img, path, nrow=nrow, padding=padding, pad_value=pad_value, normalize=normalize, value_range=value_range)
-    return
-
-
-def to_device(input, device):
-    output = recur(lambda x, y: x.to(y), input, device)
-    return output
-
-
-def ntuple(n):
-    def parse(x):
-        if isinstance(x, container_abcs.Iterable) and not isinstance(x, str):
-            return x
-        return tuple(repeat(x, n))
-
-    return parse
-
-
-def apply_fn(module, fn):
-    for n, m in module.named_children():
-        if hasattr(m, fn):
-            exec('m.{0}()'.format(fn))
-        if sum(1 for _ in m.named_children()) != 0:
-            exec('apply_fn(m,\'{0}\')'.format(fn))
-    return
-
-
-def recur(fn, input, *args):
-    if isinstance(input, torch.Tensor) or isinstance(input, np.ndarray):
-        output = fn(input, *args)
-    elif isinstance(input, list):
-        output = []
-        for i in range(len(input)):
-            output.append(recur(fn, input[i], *args))
-    elif isinstance(input, tuple):
-        output = []
-        for i in range(len(input)):
-            output.append(recur(fn, input[i], *args))
-        output = tuple(output)
-    elif isinstance(input, dict):
-        output = {}
-        for key in input:
-            output[key] = recur(fn, input[key], *args)
-    elif isinstance(input, str):
-        output = input
-    elif input is None:
-        output = None
-    else:
-        raise ValueError('Not valid input type')
-    return output
-
-
-def process_dataset(dataset):
-    cfg['data_size'] = {'train': len(dataset['train']), 'test': len(dataset['test'])}
-    cfg['target_size'] = dataset['train'].target_size
-    return
-
-
-def process_control():
-    data_shape = {'MNIST': [1, 28, 28], 'FashionMNIST': [1, 28, 28], 'SVHN': [3, 32, 32], 'CIFAR10': [3, 32, 32],
-                  'CIFAR100': [3, 32, 32]}
-    cfg['linear'] = {}
-    cfg['mlp'] = {'hidden_size': 128, 'scale_factor': 2, 'num_layers': 2, 'activation': 'relu'}
-    cfg['cnn'] = {'hidden_size': [64, 128, 256, 512]}
-    cfg['resnet9'] = {'hidden_size': [64, 128, 256, 512]}
-    cfg['resnet18'] = {'hidden_size': [64, 128, 256, 512]}
-    cfg['wresnet28x2'] = {'depth': 28, 'widen_factor': 2, 'drop_rate': 0.0}
-    cfg['wresnet28x8'] = {'depth': 28, 'widen_factor': 8, 'drop_rate': 0.0}
-    cfg['data_name'] = cfg['control']['data_name']
-    cfg['data_shape'] = data_shape[cfg['data_name']]
-    cfg['model_name'] = cfg['control']['model_name']
-    model_name = cfg['model_name']
-    cfg[model_name]['shuffle'] = {'train': True, 'test': False}
-    cfg[model_name]['optimizer_name'] = 'SGD'
-    cfg[model_name]['lr'] = 1e-1
-    cfg[model_name]['momentum'] = 0.9
-    cfg[model_name]['weight_decay'] = 5e-4
-    cfg[model_name]['nesterov'] = True
-    cfg[model_name]['scheduler_name'] = 'CosineAnnealingLR'
-    cfg[model_name]['num_epochs'] = 400
-    cfg[model_name]['batch_size'] = {'train': 250, 'test': 250}
-    return
-
-
-def make_stats():
-    stats = {}
-    stats_path = os.path.join('res', 'stats')
-    makedir_exist_ok(stats_path)
-    filenames = os.listdir(stats_path)
-    for filename in filenames:
-        stats_name = os.path.splitext(filename)[0]
-        stats[stats_name] = load(os.path.join(stats_path, filename))
-    return stats
-
-
-class Stats(object):
-    def __init__(self, dim):
-        self.dim = dim
-        self.n_samples = 0
-        self.n_features = None
-        self.mean = None
-        self.std = None
-
-    def update(self, data):
-        data = data.transpose(self.dim, -1).reshape(-1, data.size(self.dim))
-        if self.n_samples == 0:
-            self.n_samples = data.size(0)
-            self.n_features = data.size(1)
-            self.mean = data.mean(dim=0)
-            self.std = data.std(dim=0)
-        else:
-            m = float(self.n_samples)
-            n = data.size(0)
-            new_mean = data.mean(dim=0)
-            new_std = 0 if n == 1 else data.std(dim=0)
-            old_mean = self.mean
-            old_std = self.std
-            self.mean = m / (m + n) * old_mean + n / (m + n) * new_mean
-            self.std = torch.sqrt(m / (m + n) * old_std ** 2 + n / (m + n) * new_std ** 2 + m * n / (m + n) ** 2 * (
-                    old_mean - new_mean) ** 2)
-            self.n_samples += n
-        return
-
-
-def make_optimizer(parameters, tag):
-    if cfg[tag]['optimizer_name'] == 'SGD':
-        optimizer = optim.SGD(parameters, lr=cfg[tag]['lr'], momentum=cfg[tag]['momentum'],
-                              weight_decay=cfg[tag]['weight_decay'], nesterov=cfg[tag]['nesterov'])
-    elif cfg[tag]['optimizer_name'] == 'Adam':
-        optimizer = optim.Adam(parameters, lr=cfg[tag]['lr'], betas=cfg[tag]['betas'],
-                               weight_decay=cfg[tag]['weight_decay'])
-    elif cfg[tag]['optimizer_name'] == 'LBFGS':
-        optimizer = optim.LBFGS(parameters, lr=cfg[tag]['lr'])
-    else:
-        raise ValueError('Not valid optimizer name')
-    return optimizer
-
-
-def make_scheduler(optimizer, tag):
-    if cfg[tag]['scheduler_name'] == 'None':
-        scheduler = optim.lr_scheduler.MultiStepLR(optimizer, milestones=[65535])
-    elif cfg[tag]['scheduler_name'] == 'StepLR':
-        scheduler = optim.lr_scheduler.StepLR(optimizer, step_size=cfg[tag]['step_size'], gamma=cfg[tag]['factor'])
-    elif cfg[tag]['scheduler_name'] == 'MultiStepLR':
-        scheduler = optim.lr_scheduler.MultiStepLR(optimizer, milestones=cfg[tag]['milestones'],
-                                                   gamma=cfg[tag]['factor'])
-    elif cfg[tag]['scheduler_name'] == 'ExponentialLR':
-        scheduler = torch.optim.lr_scheduler.ExponentialLR(optimizer, gamma=0.99)
-    elif cfg[tag]['scheduler_name'] == 'CosineAnnealingLR':
-        scheduler = optim.lr_scheduler.CosineAnnealingLR(optimizer, T_max=cfg[tag]['num_epochs'], eta_min=0)
-    elif cfg[tag]['scheduler_name'] == 'ReduceLROnPlateau':
-        scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer, mode='min', factor=cfg[tag]['factor'],
-                                                         patience=cfg[tag]['patience'], verbose=False,
-                                                         threshold=cfg[tag]['threshold'], threshold_mode='rel',
-                                                         min_lr=cfg[tag]['min_lr'])
-    elif cfg[tag]['scheduler_name'] == 'CyclicLR':
-        scheduler = optim.lr_scheduler.CyclicLR(optimizer, base_lr=cfg[tag]['lr'], max_lr=10 * cfg[tag]['lr'])
-    else:
-        raise ValueError('Not valid scheduler name')
-    return scheduler
-
-
-def resume(path, verbose=True, resume_mode=1):
-    if os.path.exists(path) and resume_mode == 1:
-        result = load(path)
-        if verbose:
-            print('Resume from {}'.format(result['epoch']))
-    else:
-        if resume_mode == 1:
-            print('Not exists: {}'.format(path))
-        result = None
-    return result
-
-
-def collate(input):
-    for k in input:
-        input[k] = torch.stack(input[k], 0)
-    return input
